% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acum_p.R
\name{acum_p}
\alias{acum_p}
\title{Accumulate a monthly percentage variation series in n periods}
\usage{
acum_p(data, n)
}
\arguments{
\item{data}{Column}

\item{n}{Period}
}
\value{
Vector
}
\description{
Accumulate a monthly percentage variation series in n periods
}
