% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_message.R
\name{am_message}
\alias{am_message}
\title{Messages in Console}
\usage{
am_message(text, silent)
}
\arguments{
\item{text}{String}

\item{silent}{Logical}
}
\value{
Message in console
}
\description{
Messages in Console
}
\examples{
\dontrun{
am_message(text = "Starting step...", silent = FALSE)
}
}
