% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_etl.R
\name{data_etl}
\alias{data_etl}
\title{ETL script}
\usage{
data_etl(silent = FALSE)
}
\arguments{
\item{silent}{Logical}
}
\value{
Tibble
}
\description{
ETL script
}
\examples{
\dontrun{
data_download <- data_etl(silent = FALSE)
}
}
