% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_split.R
\name{data_split}
\alias{data_split}
\title{Split samples}
\usage{
data_split(data, train_end = 2019, silent = FALSE)
}
\arguments{
\item{data}{Data input}

\item{train_end}{Numeric date (YYYY) to split samples}

\item{silent}{Logical}
}
\value{
List
}
\description{
Split samples
}
\examples{
\dontrun{
silent = FALSE

# ETL etep
data_download <- data_etl(silent = silent)

# Data split step (train and test)
data_samples <- data_split(data_download, train_end = 2019, silent = silent)
}
}
