% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_data_gen}
\alias{fast_data_gen}
\title{Simulate data from linear regression model}
\usage{
fast_data_gen(n, p, R2, beta_nonzero)
}
\arguments{
\item{n}{an integer represents the sample size.}

\item{p}{an integer which is the dimension of covariates.}

\item{R2}{a number between 0 and 1 to indicate the signal to noise ratio.}

\item{beta_nonzero}{a vector of q (q > 0) non-zero coefficients for the first q covariates.}
}
\value{
a list of four variables
\describe{
 \item{x}{an n by p matrix, n observations of p-dimensional covariates}
 \item{y}{an n by 1 vector, n observations of response variable}
 \item{sigma2}{a positive number for noise variance}
 \item{beta_nonzero}{the true non-zero regression coefficients for the first q covariates}
}
}
\description{
Simulate data from linear regression model
}
\examples{
dat <- fast_data_gen(1000,10000,0.9,c(1,2,3))
}
\author{
Jian Kang \email{<jiankang@umich.edu>}
}
