% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_PMS_cpp}
\alias{fast_PMS_cpp}
\title{Fast posterior mean screening without prior knowledge}
\usage{
fast_PMS_cpp(x, y, theta = 0.001)
}
\arguments{
\item{x:}{n x p covariate matrix}

\item{y:}{n x 1 response vector}

\item{theta:}{positive number control the prior precision. Default value is 1e-3.}
}
\value{
a list of two variables,
\describe{
\item{pms_select}{indices of the predictors from most important to least important}
\item{pms_stat}{a p by 1 vector of posterior mean statistics}
}
}
\description{
Fast posterior mean screening without prior knowledge
}
\author{
Jian Kang \email{<jiankang@umich.edu>}
}
