% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_read_imgs_mask}
\alias{fast_read_imgs_mask}
\title{Fast Read multiple imaging data files on a brain mask}
\usage{
fast_read_imgs_mask(imgfiles, maskfile, verbose = 100L)
}
\arguments{
\item{imgfiles}{a character vector of imaging file names}

\item{maskfile}{a character scalar for the  mask file}

\item{verbose}{a positive integer showing the progress on the number of images being loaded}
}
\value{
a matrix of dimension n by p, where n is the number of images and p is the number of voxels on the mask
}
\description{
Fast Read multiple imaging data files on a brain mask
}
\examples{
maskfile <- file.path(system.file("nifti", package="PMS"),"MNI-maxprob-thr0-2mm.nii.gz")
mask <- oro.nifti::readNIfTI(maskfile)
imgfiles <- file.path(system.file("nifti", package="PMS"),sprintf("VBM_example_0\%d.nii.gz",1:5))
img_dat <- fast_read_imgs_mask(imgfiles,maskfile)
}
\author{
Jian Kang \email{<jiankang@umich.edu>}
}
