% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMS.R
\name{find_brain_image_neighbors}
\alias{find_brain_image_neighbors}
\title{Find the neighbors for each voxel in images}
\usage{
find_brain_image_neighbors(img, mask = NULL, radius = 1)
}
\arguments{
\item{img}{an nifti object.}

\item{mask}{an nifti object.
\code{mask > 0} specifies which voxels are on the mask.
Default value is NULL indicating all voxels are considered.}

\item{radius}{the size (voxel) of neighborhood.}
}
\value{
a list objects consisting of the following components
\describe{
\item{img_nb}{a matrix of dimension p by m indicating the neighbor indices of each voxel on the mask with respect to the original image,
where p is the number of voxels on the masks and m is the maximum number of neighbors across all the voxels}
\item{mask_img_nb}{a matrix of dimension p by m indicating the neighbor indices of each voxel on the mask with respect to the masked image}
\item{num_neighbors}{a vector of integers indicating the number of neighbors for each voxel}
\item{maskidx}{a vector of integers indicating the indices (1-based) of the orginal image where the mask takes a positve value}
\item{maskcoords}{a matrix of dimension p by d for coordinates}
}
}
\description{
Find the neighbors for each voxel in images
}
\examples{
maskfile <- file.path(system.file("nifti", package="PMS"),"MNI-maxprob-thr0-2mm.nii.gz")
mask <- oro.nifti::readNIfTI(maskfile)
imgfile <- file.path(system.file("nifti", package="PMS"),"VBM_example_01.nii.gz")
img <- oro.nifti::readNIfTI(imgfile)
nb <- find_brain_image_neighbors(img, mask,radius=1)
}
\author{
Jian Kang <jiankang@umich.edu>
}
