% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_accuracy.R
\name{model_accuracy}
\alias{model_accuracy}
\title{Report model performance}
\usage{
model_accuracy(data, silent = FALSE)
}
\arguments{
\item{data}{Data input}

\item{silent}{Logical}
}
\value{
Tibble
}
\description{
Report model performance
}
\examples{
\dontrun{
silent = FALSE

# ETL etep
data_download <- data_etl(silent = silent)

# Data split step (train and test)
data_samples <- data_split(data_download, train_end = 2019, silent = silent)

# Fit models
data_fit <- model_fit(data_samples, silent = silent)

# Model performance step (in sample)
data_performance <- model_accuracy(data_fit, silent = silent)
}
}
