% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{model_fit}
\alias{model_fit}
\title{Fit models}
\usage{
model_fit(data, silent = FALSE)
}
\arguments{
\item{data}{Data input}

\item{silent}{Logical}
}
\value{
List
}
\description{
Fit models
}
\examples{
\dontrun{
silent = FALSE

# ETL etep
data_download <- data_etl(silent = silent)

# Data split step (train and test)
data_samples <- data_split(data_download, train_end = 2019, silent = silent)

# Fit models
data_fit <- model_fit(data_samples, silent = silent)
}
}
