% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_forecast.R
\name{model_forecast}
\alias{model_forecast}
\title{Forecast "h" periods ahead}
\usage{
model_forecast(data_fit, data_frcst, silent = FALSE)
}
\arguments{
\item{data_fit}{Data input (model fit object)}

\item{data_frcst}{Data input (model samples object)}

\item{silent}{Logical}
}
\value{
List
}
\description{
Forecast "h" periods ahead
}
\examples{
\dontrun{
silent = FALSE

# ETL etep
data_download <- data_etl(silent = silent)

# Data split step (train and test)
data_samples <- data_split(data_download, train_end = 2019, silent = silent)

# Fit models
data_fit <- model_fit(data_samples, silent = silent)

# Forecast
data_forecast <- model_forecast(data_fit, data_samples, silent = silent)
}
}
