% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_save.R
\name{model_save}
\alias{model_save}
\title{Save model results}
\usage{
model_save(..., path = "data", silent = FALSE)
}
\arguments{
\item{...}{Objects to save}

\item{path}{Directory to save .rds file}

\item{silent}{Logical}
}
\value{
Save .rds file
}
\description{
Save model results
}
\examples{
\dontrun{
silent = FALSE

# ETL
data_download <- retry(data_etl(silent = silent))

# Data split (train and test)
data_samples <- data_split(data_download, train_end = train_end, silent = silent)

# Fit models
data_fit <- model_fit(data_samples, silent = silent)

# Model performance step (in sample)
data_performance <- model_accuracy(data_fit, silent = silent)

# Forecast
data_forecast <- model_forecast(data_fit, data_samples, silent = silent)

# Save results step
model_save(data_fit, data_performance, path = "data", silent = silent)
}
}
