% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMS.R
\name{pms_screening}
\alias{pms_screening}
\title{Posterior Mean Variable Screening}
\usage{
pms_screening(
  x,
  y,
  family = "gaussian",
  method = "spatial_fixed",
  mu = NULL,
  Lambda = NULL,
  Lambda_s = NULL,
  theta = NULL,
  C_rho = NULL,
  L = NULL,
  B = NULL,
  idx_s = NULL,
  rank_s = NULL,
  xgrid = NULL,
  selected_num = NULL,
  num_1 = NULL,
  cut_num_1 = NULL,
  leng = NULL,
  trho = NULL
)
}
\arguments{
\item{x}{a numeric matrix represents the predictor variables with each row be a p-dimensional observation.}

\item{y}{a numeric vector of response with each component be an observation.}

\item{family}{the model type whose choices are the same as glmnet.}

\item{method}{the specific construction of prior knowlwdge in PMS method. "spatial_fixed" indicates Lambda is a known matrix;
"selection", "rank" and "group" respectively stands for PMS method apply prior selected features set,
 ranking information as well as group structure as prior mean knowledge; and "spatial", "network" respectively means that prior
 covariance matrix Lambda has spatial and network structure with unknown parameters.}

\item{mu}{a numeric vector stands for known prior mean information, which can be simply set as "0" if no prior mean information is provided in methods "spatial" and "network".}

\item{Lambda}{a numeric matrix of known prior covariance knowledge. Similarly, we can choose "Lambda" as identity matrix if prior covariance knowledge is unavailable in
"selection", "rank" or "group" methods.}

\item{Lambda_s}{a numeric matrix represents the correlation matrix of coefficients in prior selected features(different groups) when apply "selection" or "group" method.}

\item{theta}{a numeric scalar corresponds to the value of \eqn{\theta} in PMS statistics.}

\item{C_rho}{a numeric 2-dimensional vector with 2 components correspond to the endpoints of the intervel appears in the distribution of parameter rho in
matrix Lambda when use "spatial" method.}

\item{L}{a numeric matrix stands for the normalized graph Laplacian matrix in Lambda under the "network" method.}

\item{B}{a numeric matrix stands for the group structure when apply "group" method.}

\item{idx_s}{a vecter corresponds to the index of prior selected features, which is only needed in "selection" and "rank" methods.}

\item{rank_s}{a vector corresponds to the ranking information of prior selected features in "rank" method.}

\item{xgrid}{a matrix represents the grids for covariates under image regression case.}

\item{selected_num}{an integer ranges from 1 to p, which represents the number of features selected.}

\item{num_1}{an integer, describes the total times of sampling during parameter estimation in "rank", "spatial" and "network" methods.}

\item{cut_num_1}{an integer, we throw away the first cut_num_1 samples when estimate parameter in "rank", "spatial" and "network" methods by sampling.}

\item{leng}{an integer stands for the sampling interval in the procedure of parameter estimation in "rank", "spatial" and "network" methods.}

\item{trho}{a numeric scalar controlling the degree of sparsity to matrix Lambda in method "spatial".}
}
\value{
a list of three variables, where "pms_select" represents the index of selected features with number of selected_num,
       "pms_abs" is the absolute value of PMS statistics and "elapsed" be the computation time.
}
\description{
A variable screening method for linear regression model with different constructions of prior knowledge based on the framework of Bayesian modelling
}
\examples{

set.seed(1000)
rs = 9
n = 200
p = 10000
theta = 1e-3
beta_nonzero = c(3,3,3,3,3,-7.5)
dat = data_gen(n,p,rs,beta_nonzero)
S_0 = c(2,6)
Lambda = Matrix::sparseMatrix(i=c(1:p),j=c(1:p),x=rep(1,p))
Lambda_s = Matrix::sparseMatrix(i=c(1:2),j=c(1:2),x=rep(1,2))
PMS_res = pms_screening(x = dat$x,y = dat$y,family = "gaussian",
method = "selection",Lambda = Lambda,Lambda_s = Lambda_s,
theta = theta,selected_num = p,idx_s = S_0)

}
\references{
Prior knowledge guided ultra-high dimensional variable screening with application to neuroimaging data,working paper.
}
\author{
Jie He \email{<jiehe@umich.edu>}, Jian Kang \email{<jiankang@umich.edu>}
}
