% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry.R
\name{retry}
\alias{retry}
\title{Retry an expression}
\usage{
retry(foo, tries = 3, silent = FALSE)
}
\arguments{
\item{foo}{The expression}

\item{tries}{Number of tries}

\item{silent}{Silent (logical)}
}
\value{
Output of expression or error
}
\description{
Retry an expression
}
\examples{
\dontrun{
retry(getwd())
}
}
