% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dashboard.R
\name{run_dashboard}
\alias{run_dashboard}
\title{Create dashboard}
\usage{
run_dashboard(..., save_at = "docs", silent = FALSE)
}
\arguments{
\item{...}{Other arguments to run_model()}

\item{save_at}{Directory to save .html file}

\item{silent}{Logical}
}
\value{
Dashboard
}
\description{
Create dashboard
}
\examples{
\dontrun{
run_dashboard()
}

}
