% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Run model}
\usage{
run_model(
  silent = FALSE,
  train_end = 2019,
  save_results = TRUE,
  path_results = "data"
)
}
\arguments{
\item{silent}{Logical}

\item{train_end}{Date to split samples}

\item{save_results}{Logical}

\item{path_results}{Directory to save results (used when save_results = TRUE)}
}
\value{
List
}
\description{
Run model
}
\examples{
\dontrun{
data_model <- run_model(save_results = FALSE)
}
}
